move title override of chests into CoreFeature
