+++ mineinabyss-core/src/main/kotlin/com/mineinabyss/mineinabyss/core/MIAConfig.kt
@@ -17,7 +17,6 @@ interface MIAConfig {
     /**
      * @param layers A list of all the layers and sections composing them to be registered.
      * @property hubSection The hub section of the abyss, a safe place for living and trading.
-     * @property guilds Guild related options.
      */
     @Serializable
     class Data(
