+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/core/CoreListener.kt
@@ -1,0 +1,5 @@
+package com.mineinabyss.core
+
+import org.bukkit.event.Listener
+
+class CoreListener : Listener
